% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_network.R
\name{confusion2network}
\alias{confusion2network}
\alias{prediction2network}
\title{Build network}
\usage{
confusion2network(cf, true_labels = NULL, pred_labels = NULL,
  true_label_prefix = NULL, pred_label_prefix = NULL, width = "100\%",
  height = "100\%", edge_threshold = 0.1, exportButton = FALSE)

prediction2network(pf, true_col, pred_col, true_labels = NULL,
  pred_labels = NULL, width = "100\%", height = "100\%",
  edge_threshold = 0.1, exportButton = FALSE)
}
