% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_input.R
\name{confusion2nodesedges}
\alias{confusion2nodesedges}
\alias{prediction2nodesedges}
\title{Convert to graph nodes}
\usage{
confusion2nodesedges(df_or_matrix, true_labels = NULL, pred_labels = NULL,
  true_label_prefix = NULL, pred_label_prefix = NULL)

prediction2nodesedges(prediction_df, true_col, pred_col, true_labels = NULL,
  pred_labels = NULL, group_cols = NULL)
}
